#1195 detect (if the ps command is available on unix) if the process is still active when using the 'ps' command in the CLI or if checking the process status for the container status; so we can detect if a process dies or is killed

+++ process/process-manager/src/main/java/io/fabric8/process/manager/ProcessController.java
@@ -34,5 +34,5 @@ public interface ProcessController extends Serializable {
     int status() throws Exception;
     int configure() throws Exception;
 
-    Integer getPid() throws IOException;
+    Long getPid() throws IOException;
 }
